足踏み追加＆マップ大幅更新。UIも

+++ ローグライク/meshfield.cpp
@@ -70,7 +70,7 @@ void CMeshField::MeshField_Create(int tex, int width, int height, int wcube, int
 				for (int x = 0; x <= g_mesh[i].widthcube; x++)
 				{
 					// バーテックス
-					pV[y * g_mesh[i].widthcube + x + y].position = D3DXVECTOR3(x * (g_mesh[i].widthmax / g_mesh[i].widthcube) - g_mesh[i].widthmax / 2, g_mesh[i].pos.y, g_mesh[i].heightmax / 2 - (g_mesh[i].heightmax / g_mesh[i].heightcube) * y);
+					pV[y * g_mesh[i].widthcube + x + y].position = D3DXVECTOR3((float)(x * (g_mesh[i].widthmax / g_mesh[i].widthcube) - g_mesh[i].widthmax / 2), (float)g_mesh[i].pos.y, (float)(g_mesh[i].heightmax / 2 - (g_mesh[i].heightmax / g_mesh[i].heightcube) * y));
 					pV[y * g_mesh[i].widthcube + x + y].normal = D3DXVECTOR3(0.0f, 1.0f, 0.0f);
 					pV[y * g_mesh[i].widthcube + x + y].color = D3DCOLOR_RGBA(255, 255, 255, 255);
 					pV[y * g_mesh[i].widthcube + x + y].texcoord = D3DXVECTOR2(1.0f * (x % 2), 1.0f * (y % 2));
